clean up + rename
