+++ Mirage-Android-WebView-Plugin/app/src/main/java/com/mirage/webviewplugin/CustomJavaScriptInterface.java
@@ -1,4 +1,4 @@
-package com.mirage.miragewebviewandroidplugin;
+package com.mirage.webviewplugin;
 
 import android.webkit.JavascriptInterface;
