+++ Mirage-Android-WebView-Plugin/app/src/main/java/com/mirage/webviewplugin/CustomWebViewClient.java
@@ -1,7 +1,6 @@
-package com.mirage.miragewebviewandroidplugin;
+package com.mirage.webviewplugin;
 
 import android.graphics.Bitmap;
-import android.graphics.Color;
 import android.view.View;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
