+++ Mirage-Android-WebView-Plugin/app/src/main/java/com/mirage/webviewplugin/UnityBridge.java
@@ -1,4 +1,4 @@
-package com.mirage.miragewebviewandroidplugin;
+package com.mirage.webviewplugin;
 
 import android.content.Context;
 import android.content.Intent;
